+++ pkg/grouper/grouper.go
@@ -3,10 +3,10 @@ package grouper
 import (
 	"slices"
 	"sort"
-	"strings"
 
 	"golang.org/x/exp/maps"
 
+	"github.com/google/osv-scanner/internal/identifiers"
 	"github.com/google/osv-scanner/pkg/models"
 )
 
@@ -19,26 +19,6 @@ func hasAliasIntersection(v1, v2 IDAliases) bool {
 	}
 	// Check if either IDs are in the others' aliases.
 	return slices.Contains(v1.Aliases, v2.ID) || slices.Contains(v2.Aliases, v1.ID)
-}
-
-func sortIDs(groupIDs []string) []string {
-	var otherIDs []string
-	DSA := ""
-	for _, id := range groupIDs {
-		if strings.Split(id, "-")[0] == "DSA" {
-			// DSA only has CVEs in the same group, not other DSAs.
-			DSA = id
-		} else {
-			otherIDs = append(otherIDs, id)
-		}
-	}
-
-	sort.Strings(otherIDs)
-	if DSA == "" {
-		return otherIDs
-	}
-
-	return append([]string{DSA}, otherIDs...)
 }
 
 // Group groups vulnerabilities by aliases.
@@ -77,7 +57,7 @@ func Group(vulns []IDAliases) []models.GroupInfo {
 	result := make([]models.GroupInfo, 0, len(sortedKeys))
 	for _, key := range sortedKeys {
 		// Sort the strings so they are always in the same order
-		extractedGroups[key] = sortIDs(extractedGroups[key])
+		slices.SortFunc(extractedGroups[key], identifiers.IDSortFunc)
 
 		// Add IDs to aliases
 		extractedAliases[key] = append(extractedAliases[key], extractedGroups[key]...)
